# perf: refactor host:port extraction method to make it zero-allocation