--- a/internal/netutil/netutil.go
+++ b/internal/netutil/netutil.go
@@ -18,7 +18,6 @@
 package netutil
 
 import (
-	"net"
 	"net/http"
 	"net/netip"
 	"strconv"
@@ -54,11 +53,11 @@ func parseForwardedHeader(header http.Header) (netip.Addr, uint16) {
 		return netip.Addr{}, 0
 	}
 
-	return ParseIPPort(MaybeSplitHostPort(forwarded.For))
+	return SplitAddrPort(forwarded.For)
 }
 
 func parseXRealIP(header http.Header) (netip.Addr, uint16) {
-	return ParseIPPort(MaybeSplitHostPort(getHeader(header, "X-Real-Ip", "x-real-ip")))
+	return SplitAddrPort(getHeader(header, "X-Real-Ip", "x-real-ip"))
 }
 
 func parseXForwardedFor(header http.Header) (netip.Addr, uint16) {
@@ -66,7 +65,7 @@ func parseXForwardedFor(header http.Header) (netip.Addr, uint16) {
 		if sep := strings.IndexRune(xff, ','); sep > 0 {
 			xff = xff[:sep]
 		}
-		return ParseIPPort(MaybeSplitHostPort(strings.TrimSpace(xff)))
+		return SplitAddrPort(strings.TrimSpace(xff))
 	}
 	return netip.Addr{}, 0
 }
@@ -124,50 +123,64 @@ func parseForwarded(f string) forwardedHeader {
 				continue
 			}
 		}
-		switch strings.ToLower(key) {
-		case "for":
+		switch {
+		case strings.EqualFold(key, "for"):
 			result.For = value
-		case "host":
+		case strings.EqualFold(key, "host"):
 			result.Host = value
-		case "proto":
+		case strings.EqualFold(key, "proto"):
 			result.Proto = value
 		}
 	}
 	return result
 }
 
-// ParseIPPort parses h as an IP and, if successful and p is non-empty, p as a port.
-// If h cannot be parsed as an IP or p is non-empty and cannot be parsed as a port,
-// ParseIPPort will return (nil, 0). If p is empty, 0 will be returned for the port.
-func ParseIPPort(h, p string) (netip.Addr, uint16) {
-	ip, err := netip.ParseAddr(h)
-	if err != nil {
+// SplitAddrPort splits a network address of the form "host",
+// "host:port", "[host]:port" or "[host]:port" into a netip.Addr
+// and port.
+//
+// If input has no port, 0 will be returned for the port.
+// If input cannot be parsed or it is empty, (invalidip, 0) will
+// be returned.
+func SplitAddrPort(in string) (netip.Addr, uint16) {
+	if in == "" {
 		return netip.Addr{}, 0
 	}
-	if p == "" {
-		return ip, 0
-	}
-	port, err := strconv.ParseUint(p, 10, 16)
-	if err != nil {
-		return netip.Addr{}, 0
-	}
-	return ip, uint16(port)
-}
 
-// MaybeSplitHostPort returns the result of net.SplitHostPort if it
-// would return without an error, otherwise it returns (in, ""). This
-// can be used when splitting a string which may be either a host, or
-// host:port pair.
-func MaybeSplitHostPort(in string) (host, port string) {
-	if strings.LastIndexByte(in, ':') == -1 {
-		// In the common (relative to other "errors") case that
-		// there is no colon, we can avoid allocations by not
-		// calling SplitHostPort.
-		return in, ""
+	// IPv6
+	if strings.Count(in, ":") > 1 {
+		// If [ is missing, there is no port
+		if strings.Index(in, "[") == -1 {
+			if addr, err := netip.ParseAddr(in); err == nil {
+				return addr, 0
+			}
+
+			return netip.Addr{}, 0
+		}
+
+		// [host]:port
+		if addr, err := netip.ParseAddrPort(in); err == nil {
+			return addr.Addr(), addr.Port()
+		}
+
+		return netip.Addr{}, 0
 	}
-	host, port, err := net.SplitHostPort(in)
-	if err != nil {
-		return in, ""
+
+	// IPv4
+
+	// If : is missing, there is no port
+	if strings.Index(in, ":") == -1 {
+		if addr, err := netip.ParseAddr(in); err == nil {
+			return addr, 0
+		}
+
+		return netip.Addr{}, 0
 	}
-	return host, port
+
+	// host:port
+	if addr, err := netip.ParseAddrPort(in); err == nil {
+		return addr.Addr(), addr.Port()
+	}
+
+	return netip.Addr{}, 0
 }