<!-- Thanks for sending a pull request!

If this is your first contribution, please review and sign our contributor agreement -
https://www.elastic.co/contributor-agreement.

Guidelines:
 - Prefer small PRs, and split changes into multiple logical commits where they must
   be delivered in a single PR.
 - If the PR is incomplete and not yet ready for review, open it as a Draft.
 - Once the PR is marked ready for review it is expected to pass all tests and linting,
   and you should not force-push any changes.

See also https://github.com/elastic/apm-server/blob/main/CONTRIBUTING.md for more tips on contributing.
-->

## Motivation/summary

Replace ParseIP+MaybeSplitHostPort with SplitAddrPort and make use of
the new netip package.
Avoid using net.SplitHostPort as it creates a string allocation.
Use EqualFold when parsing the Forwarded header to avoid lowercasing
the string and creating additional allocations.

Benchmarks:

```
name                                  old time/op    new time/op    delta
ContextReset/Forwarded_ipv4-16          2.47µs ± 2%    2.52µs ± 3%    +1.98%  (p=0.022 n=10+10)
ContextReset/Forwarded_ipv6-16          2.58µs ± 3%    2.54µs ± 1%    -1.22%  (p=0.020 n=10+10)
ContextReset/X-Real-IP_ipv4-16          2.47µs ± 2%    2.43µs ± 2%    -1.42%  (p=0.035 n=10+10)
ContextReset/X-Real-IP_ipv6-16          2.77µs ± 3%    2.49µs ± 2%    -9.93%  (p=0.000 n=10+10)
ContextReset/Remote_Addr_ipv4-16        2.63µs ± 4%    2.45µs ± 1%    -6.90%  (p=0.000 n=10+10)
ContextReset/Remote_Addr_ipv6-16        2.89µs ± 7%    2.46µs ± 1%   -14.78%  (p=0.000 n=10+10)
ContextReset/X-Forwarded-For_ipv4-16    2.73µs ± 4%    2.48µs ± 3%    -9.18%  (p=0.000 n=10+10)
ContextReset/X-Forwarded-For_ipv6-16    2.95µs ± 3%    2.52µs ± 2%   -14.47%  (p=0.000 n=10+9)

name                                  old alloc/op   new alloc/op   delta
ContextReset/Forwarded_ipv4-16           0.00B          0.00B           ~     (all equal)
ContextReset/Forwarded_ipv6-16           3.00B ± 0%     0.00B       -100.00%  (p=0.000 n=10+10)
ContextReset/X-Real-IP_ipv4-16           0.00B          0.00B           ~     (all equal)
ContextReset/X-Real-IP_ipv6-16           32.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
ContextReset/Remote_Addr_ipv4-16         48.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
ContextReset/Remote_Addr_ipv6-16         80.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
ContextReset/X-Forwarded-For_ipv4-16     48.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
ContextReset/X-Forwarded-For_ipv6-16     80.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)

name                                  old allocs/op  new allocs/op  delta
ContextReset/Forwarded_ipv4-16            0.00           0.00           ~     (all equal)
ContextReset/Forwarded_ipv6-16            1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
ContextReset/X-Real-IP_ipv4-16            0.00           0.00           ~     (all equal)
ContextReset/X-Real-IP_ipv6-16            1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
ContextReset/Remote_Addr_ipv4-16          1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
ContextReset/Remote_Addr_ipv6-16          2.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
ContextReset/X-Forwarded-For_ipv4-16      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
ContextReset/X-Forwarded-For_ipv6-16      2.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
```

## Checklist

<!--
Delete irrelevant items. The changelog should only be updated for user-facing changes.
Once the PR is ready for review there should be no unticked boxes.
-->

- [ ] Update [CHANGELOG.asciidoc](https://github.com/elastic/apm-server/blob/main/CHANGELOG.asciidoc)
- [ ] Update [package changelog.yml](https://github.com/elastic/apm-server/blob/main/apmpackage/apm/changelog.yml) (only if changes to `apmpackage` have been made)
- [ ] Documentation has been updated

For functional changes, consider:
- Is it observable through the addition of either **logging** or **metrics**?
- Is its use being published in **telemetry** to enable product improvement?
- Have system tests been added to avoid regression?

## How to test these changes

<!--
Explain how this PR can be tested by the reviewer: commands, dependencies, steps, etc.
If it is self-explanatory, delete this section.
-->

## Related issues

<!--
Reference the related issue(s), and make use of magic keywords where it makes sense
https://help.github.com/articles/closing-issues-using-keywords/.
-->